Change to baseCcene and socket connection to method
